stock_bot/requirements/data_pipeline_requirements.txt
alembic==1.7.5
anyio==3.4.0
apache-airflow==2.2.2
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Babel==2.9.1
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
blinker==1.4
cattrs==1.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.15
cryptography==36.0.1
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
filelock==3.4.0
finance-datareader==0.9.31
Flask==1.1.4
Flask-AppBuilder==3.4.1
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.19.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
inflection==0.5.1
iso8601==1.0.2
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.1.0
jsonschema==3.2.0
lazy-object-proxy==1.7.1
lockfile==0.12.2
lxml==4.7.1
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==1.1.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
numpy==1.21.5
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==21.3
pandas==1.3.5
pendulum==2.1.2
platformdirs==2.4.0
prison==0.2.1
psutil==5.8.0
pycparser==2.21
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
requests==2.26.0
requests-file==1.5.1
rfc3986==1.5.0
rich==10.16.1
scikit-learn==1.0.1
scipy==1.7.3
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.0
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.0.0
tqdm==4.62.3
unicodecsv==0.14.1
urllib3==1.26.7
virtualenv==20.10.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.6.0
